requirements.txt
openai
transformers
torch
haversine
pandas
fastapi
sqlmodel
sqlalchemy
psycopg2-binary
uvicorn[standard]
passlib[bcrypt]
PyJWT
python-dotenv
python-jose